# perf(matrices): use Bareiss algorithm in DomainMatrix.det