@oscarbenjamin 2021-05-23
There are further things to do here but I think that they can be left for a future PR.

What needs doing is:

1. Use pivoting for performance rather than just to avoid zeros. Pivots should be chosen to be simple i.e. ideally from the ground domain if possible. There is a bit of work on this in #20483 but it needs to be extended more generally to something like `find_reasonable_pivot`.
2. Add a sparse implementation for SDM.

In pivoting for performance consider something like the example below. Taking the `0,0` point as a pivot means that in the subsequent step every element of the matrix is multiplied by a big polynomial and all subsequent operations take place with big polynomials. Reversing the rows and columns gives something much faster. I'm not sure what the best approach here is considering that we can swap both rows and columns.
```python
In [46]: M = randMatrix(10)

In [47]: M[0,0] = random_poly(x, 10, 1, 10)

In [48]: M
Out[48]: 
⎡   10       9      8      7      6      5      4      3      2                                              ⎤
⎢4⋅x   + 10⋅x  + 2⋅x  + 3⋅x  + 7⋅x  + 7⋅x  + 6⋅x  + 8⋅x  + 2⋅x  + 2⋅x + 8  52  51  34  89  86  57  97  52  61⎥
⎢                                                                                                            ⎥
⎢                                   56                                     5   18  22  55  20  93  20  7   98⎥
⎢                                                                                                            ⎥
⎢                                   75                                     72  84  61  6   85  18  34  19  89⎥
⎢                                                                                                            ⎥
⎢                                   40                                     49  71  15  3   11  5   2   49  12⎥
⎢                                                                                                            ⎥
⎢                                   92                                     17  20  20  67  59  91  62  47  27⎥
⎢                                                                                                            ⎥
⎢                                   21                                     26  51  42  13  89  46  16  49  88⎥
⎢                                                                                                            ⎥
⎢                                   20                                     37  15  1   79  19  33  0   29  97⎥
⎢                                                                                                            ⎥
⎢                                   65                                     75  60  31  80  60  63  6   3   20⎥
⎢                                                                                                            ⎥
⎢                                   94                                     36  30  70  73  97  95  22  38  57⎥
⎢                                                                                                            ⎥
⎣                                   97                                     79  62  41  62  26  65  11  41  97⎦

In [49]: from sympy.polys.matrices import DomainMatrix

In [50]: dM = DomainMatrix.from_Matrix(M)

In [51]: %time dM.det()
CPU times: user 391 ms, sys: 3.2 ms, total: 394 ms
Wall time: 396 ms
Out[51]: 60173929000131524*x**10 + 150434822500328810*x**9 + 30086964500065762*x**8 + 45130446750098643*x**7 + 105304375750230167*x**6 + 105304375750230167*x**5 + 90260893500197286*x**4 + 120347858000263048*x**3 + 30086964500065762*x**2 + 30086964500065762*x - 1963941761526605897

In [52]: %time dM[::-1,::-1].det()
CPU times: user 9.22 ms, sys: 65 µs, total: 9.28 ms
Wall time: 9.41 ms
Out[52]: 60173929000131524*x**10 + 150434822500328810*x**9 + 30086964500065762*x**8 + 45130446750098643*x**7 + 105304375750230167*x**6 + 105304375750230167*x**5 + 90260893500197286*x**4 + 120347858000263048*x**3 + 30086964500065762*x**2 + 30086964500065762*x - 1963941761526605897
```